Make AppLifecycleService independent of http handlers
